split usechat orchestration into chatstream and drop section breaks
